Migrating to standard DI startup pattern
