+++ BudgetAnalyser/App.xaml.cs
@@ -1,9 +1,12 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
+using System.Reflection;
 using System.Text;
 using System.Windows;
 using System.Windows.Markup;
 using System.Windows.Threading;
+using BudgetAnalyser.Encryption;
 using BudgetAnalyser.Engine;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -19,6 +22,7 @@ public partial class App
     private IHost host = null!;
     private ILogger logger = null!;
     private ShellController? shellController;
+    private Stopwatch stopwatch = Stopwatch.StartNew();
 
     protected override void OnStartup(StartupEventArgs e)
     {
@@ -32,6 +36,9 @@ public partial class App
 
         Current.Exit += OnApplicationExit;
 
+        var encryptionAssembly = typeof(IFileEncryptor).GetTypeInfo().Assembly;
+        var thisAssembly = typeof(CompositionRoot).GetTypeInfo().Assembly;
+
         this.host = Host.CreateDefaultBuilder()
             .UseDefaultServiceProvider((_, options) =>
             {
@@ -40,25 +47,38 @@ public partial class App
                 options.ValidateScopes = true;
 #endif
             })
-            .ConfigureServices((_, services) => CompositionRoot.ConfigureServices(services))
+            .ConfigureServices((_, services) =>
+            {
+                services.AddEngineRegistrations();
+                services.AddAutoRegistrations(encryptionAssembly);
+                services.AddAutoRegistrations(thisAssembly);
+                services.AddReesWpfRegistrations();
+                services.AddBudgetAnalyserRegistrations();
+            })
             .Build();
 
         this.host.StartAsync().GetAwaiter().GetResult();
 
+        this.stopwatch.Stop();
+        var registrationTime = this.stopwatch.ElapsedMilliseconds;
+        this.stopwatch = Stopwatch.StartNew();
+
         this.logger = this.host.Services.GetRequiredService<ILogger>();
         this.logger.LogLevelFilter = LogLevel.Info; // hardcoded default log level.
+        this.logger.LogAlways(_ => "=========== Budget Analyser is Starting ===========");
 
         CompositionRoot.BuildApplicationObjectGraph(this.host.Services);
         this.shellController = this.host.Services.GetRequiredService<ShellController>();
-
-        this.logger.LogAlways(_ => "=========== Budget Analyser is Starting ===========");
         this.logger.LogAlways(_ => this.shellController.TopDashboardController.VersionString);
         this.shellController.Initialize();
 
-        var topLevelWindow = this.host.Services.GetRequiredService<ShellWindow>();
-        topLevelWindow.DataContext = this.shellController;
+        var topLevelWindow = new ShellWindow { DataContext = this.shellController };
         this.logger.LogInfo(_ => "Initialisation finished.");
         topLevelWindow.Show();
+
+        this.stopwatch.Stop();
+        var initialisationTime = this.stopwatch.ElapsedMilliseconds;
+        this.logger.LogAlways(_ => $"Registration took {registrationTime:N0}ms. Initialisation took {initialisationTime:N0}ms.");
     }
 
     [SuppressMessage("Microsoft.Design", "CA1031:DoNotCatchGeneralExceptionTypes",
